+++ lib/strong_migrations/database_tasks.rb
@@ -3,7 +3,7 @@ module StrongMigrations
     def migrate
       super
     rescue => e
-      if e.cause.is_a?(StrongMigrations::UnsafeMigration)
+      if e.cause.is_a?(StrongMigrations::Error)
         # strip cause and clean backtrace
         def e.cause
           nil