+++ lib/strong_migrations/migration_helpers.rb
@@ -1,14 +1,6 @@
 module StrongMigrations
   module MigrationHelpers
-
-    # Adds a new foreign key with minimal impact on concurrent updates.
-    #
-    # Example:
-    #
-    #     add_foreign_key_concurrently :articles, :authors
-    #
-    # Refer to Rails' `add_foreign_key` for more info on available options.
-    def add_foreign_key_concurrently(from_table, to_table, options = {})
+    def add_foreign_key_safely(from_table, to_table, **options)
       ensure_postgresql(__method__)
       ensure_not_in_transaction(__method__)
 
@@ -16,30 +8,25 @@ module StrongMigrations
         add_foreign_key(from_table, to_table, options.merge(validate: false))
         validate_foreign_key(from_table, to_table)
       else
-        options = foreign_key_options(from_table, to_table, options)
+        reversible do |dir|
+          dir.up do
+            options = foreign_key_options(from_table, to_table, options)
+            fk_name, column, primary_key = options.values_at(:name, :column, :primary_key)
+            primary_key ||= "id"
 
-        safety_assured do
-          fk_name, column, primary_key = options.values_at(:name, :column, :primary_key)
-          primary_key ||= "id"
+            statement = ["ALTER TABLE %s ADD CONSTRAINT %s FOREIGN KEY (%s) REFERENCES %s (%s)"]
+            statement << on_delete_update_statement(:delete, options[:on_delete]) if options[:on_delete]
+            statement << on_delete_update_statement(:update, options[:on_update]) if options[:on_update]
+            statement << "NOT VALID"
 
-          reversible do |dir|
-            dir.up do
-              execute quote_identifiers(<<~SQL, [from_table, fk_name, column, to_table, primary_key])
-                ALTER TABLE %s
-                ADD CONSTRAINT %s
-                FOREIGN KEY (%s)
-                REFERENCES %s (%s)
-                #{on_delete_update_statement(:delete, options[:on_delete])}
-                #{on_delete_update_statement(:update, options[:on_update])}
-                NOT VALID;
-              SQL
-
-              execute quote_identifiers("ALTER TABLE %s VALIDATE CONSTRAINT %s;", [from_table, fk_name])
+            safety_assured do
+              execute quote_identifiers(statement.join(" "), [from_table, fk_name, column, to_table, primary_key])
+              execute quote_identifiers("ALTER TABLE %s VALIDATE CONSTRAINT %s", [from_table, fk_name])
             end
+          end
 
-            dir.down do
-              remove_foreign_key(from_table, to_table)
-            end
+          dir.down do
+            remove_foreign_key(from_table, to_table)
           end
         end
       end
@@ -48,7 +35,7 @@ module StrongMigrations
     private
 
     def ensure_postgresql(method_name)
-      raise "`#{method_name}` is intended for Postgres usage only" unless postgresql?
+      raise StrongMigrations::Error, "`#{method_name}` is intended for Postgres only" unless postgresql?
     end
 
     def postgresql?
@@ -56,24 +43,24 @@ module StrongMigrations
     end
 
     def ensure_not_in_transaction(method_name)
-      if transaction_open?
-        raise <<~EOF
-          Cannot run `#{method_name}` inside a transaction.
-          To disable the transaction wrapping this migration, you can use `disable_ddl_transaction!`.
-        EOF
+      if connection.transaction_open?
+        raise StrongMigrations::Error, "Cannot run `#{method_name}` inside a transaction. Use `disable_ddl_transaction` to disable the transaction."
       end
     end
 
     def on_delete_update_statement(delete_or_update, action)
-      delete_or_update = delete_or_update.to_s
+      on = delete_or_update.to_s.upcase
 
       case action
-      when nil, ""
-        ""
       when :nullify
-        "ON #{delete_or_update.upcase} SET NULL"
+        "ON #{on} SET NULL"
+      when :cascade
+        "ON #{on} CASCADE"
+      when :restrict
+        "ON #{on} RESTRICT"
       else
-        "ON #{delete_or_update.upcase} #{action.upcase}"
+        # same error message as Active Record
+        raise "'#{action}' is not supported for :on_update or :on_delete.\nSupported values are: :nullify, :cascade, :restrict"
       end
     end