+++ lib/strong_migrations/unsafe_migration.rb
@@ -1,4 +1,0 @@
-module StrongMigrations
-  class UnsafeMigration < StandardError
-  end
-end